requirements.txt
fastapi
uvicorn
requests
numpy
pillow
opencv-python


